SuSEFirewallServices: Move SuSEFirewall{2,d}Services classes to new files
Yast2 coding standard requires each class to be on a separate file
named similar to the class that it includes. As such move
SuSEFirewall2Services and SuSEFirewalldServices classes to separate
files.
